+++ util-core/src/main/scala/com/twitter/util/ClosableOnce.scala
@@ -1,49 +1,16 @@
 package com.twitter.util
 
-import scala.util.control.NonFatal
-
 /**
  * A mixin trait to describe resources that are an idempotent [[Closable]].
  *
  * The first call to `close(Time)` triggers closing the resource with the
  * provided deadline while subsequent calls will yield the same `Future`
  * as the first invocation irrespective of the deadline provided.
+ *
+ * @see [[CloseOnce]] if you are mixing in or extending an existing [[Closable]]
+ *      to prevent any "inherits from conflicting members" compile-time issues.
  */
-trait ClosableOnce extends Closable {
-
-  // Our intrinsic lock for mutating the `closed` field
-  private[this] val closePromise: Promise[Unit] = Promise[Unit]()
-  private[this] var closed: Boolean = false
-
-  private[this] def once(): Boolean = closePromise.synchronized {
-    if (closed) {
-      false
-    } else {
-      closed = true
-      true
-    }
-  }
-
-  /**
-   * Close the resource with the given deadline exactly once. This deadline is
-   * advisory, giving the callee some leeway, for example to drain clients or
-   * finish up other tasks.
-   *
-   * @note if this method throws a synchronous exception, that exception will
-   *       be wrapped in a failed future.
-   */
-  protected def closeOnce(deadline: Time): Future[Unit]
-
-  final def close(deadline: Time): Future[Unit] = {
-    if (once()) {
-      val closeF =
-        try closeOnce(deadline)
-        catch { case NonFatal(ex) => Future.exception(ex) }
-      closePromise.become(closeF)
-    }
-    closePromise
-  }
-}
+trait ClosableOnce extends Closable with CloseOnce
 
 object ClosableOnce {